move `cupla/math` to `cupla/device/math`
Move the location of the math functions into the device directory.
