+++ include/cupla/device/math/Common.hpp
