+++ include/cupla/device/math/Trigo.hpp
@@ -21,7 +21,7 @@
 
 #pragma once
 
-#include "cupla/math/Common.hpp"
+#include "cupla/device/math/Common.hpp"
 #include "cupla/types.hpp"
 
 namespace cupla
@@ -33,9 +33,26 @@ inline namespace device
 inline namespace math
 {
 
-    //! Computes the absolute value.
-    CUPLA_UNARY_MATH_FN( abs, alpaka::math::ConceptMathAbs, Abs )
+    //! Computes the sine (measured in radians).
+    CUPLA_UNARY_MATH_FN( sin, alpaka::math::ConceptMathSin, Sin )
 
+    //! Computes the cosine (measured in radians).
+    CUPLA_UNARY_MATH_FN( cos, alpaka::math::ConceptMathCos, Cos )
+
+    //! Computes the tangent (measured in radians).
+    CUPLA_UNARY_MATH_FN( tan, alpaka::math::ConceptMathTan, Tan )
+
+    //! Computes the principal value of the arc sine.
+    CUPLA_UNARY_MATH_FN( asin, alpaka::math::ConceptMathAsin, Asin )
+
+    //! Computes the principal value of the arc cosine.
+    CUPLA_UNARY_MATH_FN( acos, alpaka::math::ConceptMathAcos, Acos )
+
+    //! Computes the principal value of the arc tangent.
+    CUPLA_UNARY_MATH_FN( atan, alpaka::math::ConceptMathAtan, Atan )
+
+    //! Computes the arc tangent of y/x using the signs of arguments to determine the correct quadrant.
+    CUPLA_BINARY_MATH_FN( atan2, alpaka::math::ConceptMathAtan2, Atan2 )
 
 } // namespace math
 } // namespace device
