+++ include/cupla/device_functions.hpp
@@ -25,4 +25,4 @@
 #include "cupla/device/Index.hpp"
 #include "cupla/device/Atomic.hpp"
 #include "cupla/device/SharedMemory.hpp"
-#include "cupla/math.hpp"
+#include "cupla/device/math.hpp"
